+++ tests/fixtures/saas/alchemer_fixtures.py
@@ -1,6 +1,7 @@
 import random
 import string
 from typing import Any, Dict, Generator
+import time
 
 import pydash
 import pytest
@@ -58,8 +59,8 @@ def alchemer_erasure_data(
 
     base_url = f"https://{alchemer_secrets['domain']}/v5"
     params = {
-        "api_key": alchemer_secrets["api_key"],
-        "api_key_secret": alchemer_secrets["api_key_secret"],
+        "api_token": alchemer_secrets["api_key"],
+        "api_token_secret": alchemer_secrets["api_key_secret"],
         "list_name": x_contactlist_name,
     }
     contactlist_url = f"{base_url}/contactlist/"
@@ -68,11 +69,12 @@ def alchemer_erasure_data(
     contactlist_id = response.json()["data"]["id"]
     contactlistcontact_url = f"{contactlist_url}{contactlist_id}/contactlistcontact"
     params = {
-        "api_key": alchemer_secrets["api_key"],
-        "api_key_secret": alchemer_secrets["api_key_secret"],
+        "api_token": alchemer_secrets["api_key"],
+        "api_token_secret": alchemer_secrets["api_key_secret"],
         "email_address": alchemer_erasure_identity_email,
     }
     response = requests.put(contactlistcontact_url, params=params)
+    time.sleep(5)
 
 
 @pytest.fixture