+++ tests/ops/integration_tests/saas/test_alchemer_task.py
@@ -15,8 +15,9 @@ class TestalchemerConnector:
         policy: Policy,
         erasure_policy_string_rewrite: Policy,
         alchemer_erasure_identity_email: str,
+        alchemer_erasure_data,
     ):
-        (erasure_results,) = await alchemer_runner.non_strict_erasure_request(
+        (_, erasure_results,) = await alchemer_runner.non_strict_erasure_request(
             access_policy=policy,
             erasure_policy=erasure_policy_string_rewrite,
             identities={"email": alchemer_erasure_identity_email},